Adding Meteor.
Contributor: Matt DeBergalis, founder of Meteor, matt@meteor.com
Source: internal headcount

I'm a founder of Meteor.  We are building a full-stack JavaScript
framework that makes it much faster and easier to write modern apps
where most of the code runs in the browser.

Meteor has 9 full-time technical people.  3 are women.  (Our bios page
is at http://www.meteor.com/about but is running a few people behind.)
Company-wide we are 7 female and 7 male.  And we work hard to promote
balance at Meteor Devshop and at the various Meteor meetups around the
world.  This is super important to us.