Update data.txt
I've had it on my to-do list for several months to add my company, Moontoast, to this great project. About a month ago, my company underwent a merger, resulting in the layoffs of 20% of the company. I was the only female engineer at this company and the only engineer to be laid off. This is an accurate headcount for the day after the merger, April 24, 2013:

0 women out of 11 if you count dev + ops. 0 out of 14 if you count QA.