Adding Coinbase.
Contributor: Nahid Samsami, Developer Advocate at Coinbase,
    nahid.samsami@gmail.com

Source: internal headcount

Initially I was on the business development team but I slowly moved
into the role of Developer Advocate and officially moved into the
engineering team in that capacity a month ago.

We have 15 full-time folks on the engineering team, and I'm currently
the sole female. I'm not sure if I should be included in your numbers,
as majority of the code I commit is front-end related to our API
documentation, plus some bug fixes related to our API. I'll leave it
to your discretion, it may be 14/14 or 14/1 depending on if you count
me.